01-develop-generative-ai-applications/03-Lab/requirements.txt
langchain
langchain-huggingface
huggingface_hub
python-dotenv
Flask